+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * general support for all bourne- and c-based shells #3175
 
 
-## 4.3.2 (unreleased)
+## 4.3.2 (2016-01-06)
 
 ### Deprecations/Breaking Changes
 * Further refine conda channels specification. To verify if the url of a channel
@@ -24,6 +24,8 @@
 * fix #4152 conda env export failure (#4175)
 * fix #3779 channel UNC path encoding errors on windows (#4190)
 * fix progress bar (#4191)
+* use context.channels instead of args.channel (#4199)
+* don't use local cached repodata for file:// urls (#4209)
 
 ### Non-User-Facing Changes
 * xfail anaconda token test if local token is found (#4124)
@@ -180,7 +182,7 @@
 * context-dependent setup.py files (#4057)
 
 
-## 4.2.14 (2017-01-05)
+## 4.2.14 (2017-01-06)
 
 ### Improvements
 * use install.rm_rf for TemporaryDirectory cleanup (#3425)
@@ -195,10 +197,11 @@
 * fix location of temporary hard links of index.json (#3975)
 * fix potential errors in multi-channel export and offline clone (#3995)
 * fix auxlib/packaging, git hashes are not limited to 7 characters (#4189)
+* fix compatibility with requests >=2.12, add pyopenssl as dependency (#4059)
 
 ### Non-User-Facing Changes
 * fix open-ended test failures relating to python 3.6 release (#4166)
-* allow args passed to cli.main() (#4193)
+* allow args passed to cli.main() (#4193, #4200, #4201)
 * test against python 3.6 (#4197)